fixed f5763d8 from https://gitee.com/haixiangw/security_certificate_manager/pulls/231 告警清理

+++ services/cert_manager_standard/cert_manager_service/main/os_dependency/sa/cm_sa.cpp
@@ -255,7 +255,7 @@ void CertManagerService::OnStart(const SystemAbilityOnDemandReason& startReason)
     if (startReason.GetId() == OnDemandReasonId::COMMON_EVENT &&
         startReason.GetName() == USER_REMOVED_EVENT) {
         struct CmContext context = { 0, INVALID_VALUE, {0} };
-        context.userId = startReason.GetExtraData().GetCode();
+        context.userId = (uint32_t)startReason.GetExtraData().GetCode();
         CM_LOG_D("user remove event, userId = %u", context.userId);
         CmDeleteProcessInfo(&context);
     }
